+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 ChangeLog
 =========
+
+added config file
+- transcription boundary threshold is parametrized